Hint reveal warning added, hint skip bug squashed
Users must get through a confirmation popup when revealing a hint that
will cost them fans.  This popup stops the timer.  Given time, we want
to rework this so it doesn't.

When skipping hints due to partial answers the system would even set
revealed hints to skipped, bringing up a button to reveal the hint
again.  This has been fixed.